コース選択をApp.Vueに移動
