+++ src/stores/assignmentStore.ts
@@ -6,6 +6,7 @@ import { Assignment, Status } from '../components/utils/Commons'
 export const useAssignmentStore = defineStore(
   'assignment',
   () => {
+    const selectedCourse = ref<string>('python_basic_101')
     const selectedAssignment = ref<Assignment>({
       id: 1,
       title: '',
@@ -43,6 +44,7 @@ export const useAssignmentStore = defineStore(
       }, 0)
     })
     return {
+      selectedCourse,
       selectedAssignment,
       getProblemScore,
       getTotalScore,
